move helper fns
