+++ mirror/mirror-server/src/functions/app/publish.function.ts
@@ -12,12 +12,12 @@ import * as semver from 'semver';
 import {newDeploymentID} from 'shared/src/mirror/ids.js';
 import {isSupportedSemverRange} from 'shared/src/mirror/is-supported-semver-range.js';
 import {storeModule} from 'shared/src/mirror/store-module.js';
-import type {CfModule} from '../cloudflare/create-worker-upload-form.js';
-import {getServerModuleMetadata} from '../cloudflare/get-server-modules.js';
-import {publish as publishToCloudflare} from '../cloudflare/publish.js';
-import {findNewestMatchingVersion} from '../find-newest-matching-version.js';
-import {appAuthorization, userAuthorization} from './validators/auth.js';
-import {validateSchema} from './validators/schema.js';
+import type {CfModule} from '../../cloudflare/create-worker-upload-form.js';
+import {getServerModuleMetadata} from '../../cloudflare/get-server-modules.js';
+import {publish as publishToCloudflare} from '../../cloudflare/publish.js';
+import {findNewestMatchingVersion} from './find-newest-matching-version.js';
+import {appAuthorization, userAuthorization} from '../validators/auth.js';
+import {validateSchema} from '../validators/schema.js';
 
 // This is the API token for reflect-server.net
 // https://dash.cloudflare.com/085f6d8eb08e5b23debfb08b21bda1eb/
